libreactor: move setup() to helpers.c and delete setup.c for the main (platform) implementation
