+++ frameworks/C/libreactor/src/helpers.c
@@ -1,9 +1,12 @@
+#define _GNU_SOURCE
+
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <unistd.h>
-#include <limits.h>
 #include <string.h>
+#include <sched.h>
+#include <sys/wait.h>
 #include <err.h>
 
 #include <dynamic.h>
@@ -75,4 +78,37 @@ void json(server_context *context, clo *json_object)
 
   (void) clo_encode(json_object, json_string, sizeof(json_string));
   write_response(&context->session->stream, json_preamble, segment_string(json_string));
+}
+
+void setup()
+{
+  int e;
+  pid_t pid;
+  cpu_set_t available_cpus, cpu;
+
+  signal(SIGPIPE, SIG_IGN);
+  CPU_ZERO(&available_cpus);
+  sched_getaffinity(0, sizeof(available_cpus), &available_cpus); // Get set of all available CPUs
+
+  for (int i = 0; i < CPU_SETSIZE; i++)
+  {
+    if (CPU_ISSET(i, &available_cpus))
+    {
+      pid = fork();
+      if (pid == -1)
+        err(1, "fork");
+
+      if (pid == 0)
+      {
+       CPU_ZERO(&cpu);
+       CPU_SET(i, &cpu);
+        e = sched_setaffinity(0, sizeof cpu, &cpu);
+        if (e == -1)
+          err(1, "sched_setaffinity");
+
+        return;
+      }
+    }
+  }
+  wait(NULL);
 }
