+++ frameworks/C/libreactor/src/helpers.h
@@ -5,4 +5,6 @@ void plaintext(server_context *context, char *response);
 
 void json(server_context *context, clo *json_object);
 
+void setup();
+
 #endif /* HELPERS_H_INCLUDED */
